move touch actions in common touch action directoly
