+++ lib/appium_lib_core/common/touch_action/multi_touch.rb
