+++ lib/appium_lib_core/device.rb
@@ -1,5 +1,5 @@
-require_relative 'device/touch_actions'
-require_relative 'device/multi_touch'
+require_relative 'common/touch_action/touch_actions'
+require_relative 'common/touch_action/multi_touch'
 require_relative 'device/screen_record'
 require_relative 'device/app_state'
 require_relative 'device/clipboard_content_type'
