+++ test/functional/ios/ios/device_test.rb
@@ -44,11 +44,11 @@ class AppiumLibCoreTest
       end
 
       def test_lock_unlock
-        @@driver.lock 5
-        assert @@driver.device_locked?
+        @@driver.lock
+        @@core.wait { assert @@driver.device_locked? }
 
         @@driver.unlock
-        assert !@@driver.device_locked?
+        @@core.wait { assert !@@driver.device_locked? }
       end
 
       def test_background_reset
